Fix /server and other means of making a connection. Should get us back to 1.1 behavior. BUG:192503
svn path=/branches/work/konvi-kde4/; revision=967701

+++ src/quickconnectdialog.cpp
@@ -109,7 +109,7 @@ void QuickConnectDialog::slotOk()
 
         emit connectClicked(Konversation::PromptToReuseConnection,
                             hostNameInput->text().trimmed(),
-                            portInput->text().toUInt(),
+                            portInput->text(),
                             passwordInput->text(),
                             nickInput->text(),
                             "",
